+++ multinode-demo/bootstrap-validator.sh
@@ -144,6 +144,7 @@ args+=(
   --no-wait-for-vote-to-start-leader
   --full-rpc-api
   --allow-private-addr
+  --no-xdp
 )
 default_arg --gossip-port 8001
 default_arg --log -