+++ scripts/run.sh
@@ -121,6 +121,7 @@ args=(
   --require-tower
   --no-wait-for-vote-to-start-leader
   --no-os-network-limits-test
+  --no-xdp
 )
 # shellcheck disable=SC2086
 agave-validator "${args[@]}" $SOLANA_RUN_SH_VALIDATOR_ARGS &